Move pvmCall implementation to pvm module
