+++ crates/cheatcodes/src/evm.rs
@@ -2,8 +2,6 @@
 
 use crate::{
     inspector::{InnerEcx, RecordDebugStepInfo},
-    pvm::PvmCheatcodeInspectorStrategyContext,
-    strategy::CheatcodeInspectorStrategy,
     BroadcastableTransaction, Cheatcode, Cheatcodes, CheatcodesExecutor, CheatsCtxt, Error, Result,
     Vm::*,
 };
@@ -1256,41 +1254,4 @@ fn set_cold_slot(ccx: &mut CheatsCtxt, target: Address, slot: U256, cold: bool) 
             storage_slot.is_cold = cold;
         }
     }
-}
-
-impl Cheatcode for pvmCall {
-    fn apply(&self, state: &mut Cheatcodes) -> Result {
-        let Self { enabled } = self;
-
-        // Switch the strategy based on PVM setting
-        if *enabled {
-            state.strategy = CheatcodeInspectorStrategy::new_pvm();
-            tracing::info!("PVM mode enabled");
-        } else {
-            // Switch back to EVM strategy
-            state.strategy = CheatcodeInspectorStrategy::new_evm();
-            tracing::info!("PVM mode disabled, using EVM");
-        }
-
-        Ok(Default::default())
-    }
-
-    fn apply_full(&self, ccx: &mut CheatsCtxt, executor: &mut dyn CheatcodesExecutor) -> Result {
-        let Self { enabled } = self;
-        if *enable {
-            let ctx = get_context(ccx.state.strategy.context.as_mut());
-            todo!("select_pvm(ctx, ccx.ecx, None)");
-            tracing::info!("PVM mode enabled");
-        } else {
-            todo!("Switch back to EVM");
-            tracing::info!("PVM mode disabled, using EVM");
-        }
-    }
-}
-
-// TODO: why it's in evm module?
-fn get_context(
-    ctx: &mut dyn CheatcodeInspectorStrategyContext,
-) -> &mut PvmCheatcodeInspectorStrategyContext {
-    ctx.as_any_mut().downcast_mut().expect("expected PvmCheatcodeInspectorStrategyContext")
 }
